+++ provision.sh
@@ -1,5 +1,6 @@
 
 export CONFIG_DIR=blockr_config
+export DEBUG=false
 export SETUP_DRIVER_NAME=prepare_node_driver.sh
 export FABRIC_CFG_PATH=./$CONFIG_DIR
 export FABRIC_PATH=$GOPATH/src/github.com/hyperledger/fabric
@@ -10,7 +11,6 @@ export WAIT_SECONDS=1
 export WITH_TLS=true
 
 distribute_conf() {
-  echo ""
   echo "----------"
   echo " Distribute configuration to Node $1"
   echo "----------"
@@ -40,7 +40,7 @@ distribute_conf() {
     ORDERER_GENERAL_TLS_ROOTCAS="ordererOrganizations/$3/orderers/$1.$3/tls/ca.crt"
   fi
 
-  cat $TEMP_CFG_PATH/core.yml.template | sed "s|PEER_ID|$1| ; s|PEER_ADDRESS|$1:7051| ; s|PEER_BOOTSTRAP|$1:7051| ; s|WITH_TLS|$WITH_TLS| ; s|PEER_CERT|$CORE_PEER_TLS_CERT_FILE| ; s|PEER_KEY|$CORE_PEER_TLS_KEY_FILE| ; s|PEER_ROOTCERT|$CORE_PEER_TLS_ROOTCERT_FILE| ; s|PEER_MSP_PATH|$CORE_PEER_MSP_PATH| ; s|PEER_MSP_ID|$2| " > $TEMP_CFG_PATH/core.yaml
+  cat $TEMP_CFG_PATH/core.yml.template | sed "s|PEER_ID|$2| ; s|PEER_ENDPOINT|$1| ; s|PEER_ADDRESS|$1:7051| ; s|PEER_BOOTSTRAP|$1:7051| ; s|WITH_TLS|$WITH_TLS| ; s|PEER_CERT|$CORE_PEER_TLS_CERT_FILE| ; s|PEER_KEY|$CORE_PEER_TLS_KEY_FILE| ; s|PEER_ROOTCERT|$CORE_PEER_TLS_ROOTCERT_FILE| ; s|PEER_MSP_PATH|$CORE_PEER_MSP_PATH| ; s|PEER_MSP_ID|$2| " > $TEMP_CFG_PATH/core.yaml
   rm $TEMP_CFG_PATH/core.yml.template
 
   cat $TEMP_CFG_PATH/orderer.yml.template | sed "s:WITH_TLS:$WITH_TLS: ; s:ORDERER_CERT:$ORDERER_GENERAL_TLS_CERTIFICATE: ; s:ORDERER_KEY:$ORDERER_GENERAL_TLS_PRIVATEKEY: ; s:ORDERER_ROOTCERT:$ORDERER_GENERAL_TLS_ROOTCAS: ; s:ORDERER_MSP_PATH:$ORDERER_MSP_PATH: ; s:ORDERER_MSP_ID:$4:   " > $TEMP_CFG_PATH/orderer.yaml
@@ -53,21 +53,70 @@ distribute_conf() {
 }
 
 prepare() {
-  echo ""
   echo "----------"
   echo " Preparing Node $1"
   echo "----------"
-  echo "Stop running daemons"
-  ssh $1 "pkill orderer"
-  ssh $1 "pkill peer"
-  echo "Remove docker images"
-  ssh $1 "docker ps -aq | xargs docker kill &> /dev/null " || echo -n "."
-  ssh $1 "docker ps -aq | xargs docker rm &> /dev/null " || echo -n "."
-  ssh $1 "docker images | grep 'dev-' | awk '{print $3}' | xargs docker rmi &> /dev/null " || echo -n "."
+
+#
+# create the driver script
+#
+  echo '#!/bin/bash' > $SETUP_DRIVER_NAME
+  echo '' >> $SETUP_DRIVER_NAME
+  echo '#----------------' >> $SETUP_DRIVER_NAME
+  echo '#' >> $SETUP_DRIVER_NAME
+  echo '# Block R Setup Driver' >> $SETUP_DRIVER_NAME
+  echo '#' >> $SETUP_DRIVER_NAME
+  echo '#----------------' >> $SETUP_DRIVER_NAME
+  echo -n 'export TARGET_CFG_PATH=' >> $SETUP_DRIVER_NAME 
+  echo $TARGET_CFG_PATH >> $SETUP_DRIVER_NAME 
+  echo -n 'export PRODUCTION_DIR=' >> $SETUP_DRIVER_NAME 
+  echo $PRODUCTION_DIR >> $SETUP_DRIVER_NAME 
+  echo 'echo " - Stop running daemons"' >> $SETUP_DRIVER_NAME
+  echo 'sudo pkill orderer' >> $SETUP_DRIVER_NAME
+  echo 'sudo pkill peer' >> $SETUP_DRIVER_NAME
+  echo 'echo " - Remove docker images"' >> $SETUP_DRIVER_NAME
+  echo 'sudo docker ps -aq | xargs docker kill &> /dev/null' >> $SETUP_DRIVER_NAME
+  echo 'sudo docker ps -aq | xargs docker rm &> /dev/null' >> $SETUP_DRIVER_NAME
+  echo "sudo docker images | grep 'dev-' | awk '{print $3}' | xargs docker rmi &> /dev/null" >> $SETUP_DRIVER_NAME
+  echo -n 'echo " - Reset configuration ' >> $SETUP_DRIVER_NAME
+  echo -n $TARGET_CFG_PATH >> $SETUP_DRIVER_NAME
+  echo '"' >> $SETUP_DRIVER_NAME
+  echo -n 'rm -rf ' >> $SETUP_DRIVER_NAME
+  echo $TARGET_CFG_PATH >> $SETUP_DRIVER_NAME
+  echo -n 'mkdir ' >> $SETUP_DRIVER_NAME
+  echo $TARGET_CFG_PATH >> $SETUP_DRIVER_NAME
+  echo 'echo " - Stop daemons"' >> $SETUP_DRIVER_NAME
+  echo 'sudo /etc/init.d/couchdb stop &> /dev/null' >> $SETUP_DRIVER_NAME
+  echo 'sudo systemctl stop kafka' >> $SETUP_DRIVER_NAME
+  if [ "$1" = "vm1" ]; then
+    echo 'sudo systemctl stop zookeeper' >> $SETUP_DRIVER_NAME
+  fi
+  echo 'echo " - Reset production repositories"' >> $SETUP_DRIVER_NAME
+  echo 'if [ -d $PRODUCTION_DIR ]; then' >> $SETUP_DRIVER_NAME
+  echo '  sudo rm -rf $PRODUCTION_DIR' >> $SETUP_DRIVER_NAME
+  echo 'fi' >> $SETUP_DRIVER_NAME
+  echo 'sudo mkdir $PRODUCTION_DIR' >> $SETUP_DRIVER_NAME
+  echo 'sudo chown $(whoami):$(whoami) $PRODUCTION_DIR' >> $SETUP_DRIVER_NAME
+  echo 'if [ -d tmp.zookeeper ]; then' >> $SETUP_DRIVER_NAME
+  echo '  sudo -rm -rf /tmp/zookeeper' >> $SETUP_DRIVER_NAME
+  echo 'fi' >> $SETUP_DRIVER_NAME
+  echo 'if [ -d tmp.kafka-logs ]; then' >> $SETUP_DRIVER_NAME
+  echo '  sudo -rm -rf /tmp/kafka-logs' >> $SETUP_DRIVER_NAME
+  echo 'fi' >> $SETUP_DRIVER_NAME
+  echo 'echo " - Start daemons"' >> $SETUP_DRIVER_NAME
+  if [ "$1" = "vm1" ]; then
+    echo 'sudo systemctl start zookeeper' >> $SETUP_DRIVER_NAME
+  fi
+  echo 'sudo systemctl start kafka' >> $SETUP_DRIVER_NAME
+  echo 'sudo /etc/init.d/couchdb start &> /dev/null' >> $SETUP_DRIVER_NAME
+
   scp -q ./$SETUP_DRIVER_NAME $1: 
   ssh $1 "chmod 777 $SETUP_DRIVER_NAME"
   ssh $1 "./$SETUP_DRIVER_NAME"
-  ssh $1 "rm ./$SETUP_DRIVER_NAME"
+  if [ "$DEBUG" != true ]; then
+    ssh $1 "rm ./$SETUP_DRIVER_NAME"
+  fi
+  rm ./$SETUP_DRIVER_NAME
 }
 
 echo ".----------------"
@@ -76,48 +125,14 @@ echo "| Block R Provisoner"
 echo "|"
 echo "'----------------"
 
-echo "----------"
-echo " Reset local configuration directory $CONFIG_DIR"
-echo "----------"
-rm -rf ./$CONFIG_DIR
-mkdir -p ./$CONFIG_DIR
-
-#
-# create the driver script
-#
-echo '#!/bin/bash' > $SETUP_DRIVER_NAME
-echo '' >> $SETUP_DRIVER_NAME
-echo '#----------------' >> $SETUP_DRIVER_NAME
-echo '#' >> $SETUP_DRIVER_NAME
-echo '# Block R Setup Driver' >> $SETUP_DRIVER_NAME
-echo '#' >> $SETUP_DRIVER_NAME
-echo '#----------------' >> $SETUP_DRIVER_NAME
-echo -n 'export TARGET_CFG_PATH=' >> $SETUP_DRIVER_NAME 
-echo $TARGET_CFG_PATH >> $SETUP_DRIVER_NAME 
-echo -n 'export PRODUCTION_DIR=' >> $SETUP_DRIVER_NAME 
-echo $PRODUCTION_DIR >> $SETUP_DRIVER_NAME 
-echo '
-echo "----------"
-echo " Reset configuration $TARGET_CFG_PATH"
-echo "----------"
-rm -rf $TARGET_CFG_PATH
-mkdir -p $TARGET_CFG_PATH
-sudo /etc/init.d/couchdb stop
-echo "----------"
-echo " Reset Production directory $PRODUCTION_DIR"
-echo "----------"
-if [ -d $PRODUCTION_DIR ]; then
-  sudo rm -rf $PRODUCTION_DIR
-fi
-sudo mkdir -p $PRODUCTION_DIR 
-sudo chown $(whoami):$(whoami) $PRODUCTION_DIR 
-sudo /etc/init.d/couchdb start
-' >> $SETUP_DRIVER_NAME 
-
 prepare vm1
 prepare vm2
-sleep $WAIT_SECONDS 
-rm ./$SETUP_DRIVER_NAME
+
+echo "----------"
+echo " Reset local configuration directory $FABRIC_CFG_PATH"
+echo "----------"
+rm -rf $FABRIC_CFG_PATH 
+mkdir -p $FABRIC_CFG_PATH 
 
 echo "----------"
 echo " Generate keys from $FABRIC_CFG_PATH/blockr-config.yaml"
@@ -159,8 +174,8 @@ if ! [ -f $FABRIC_CFG_PATH/Org2MSPanchors.tx ]; then
   exit 1
 fi
 
-distribute_conf vm1 Org1MSP nar.blockr Orderer1MSP
-distribute_conf vm2 Org2MSP car.blockr Orderer2MSP
+distribute_conf vm1 Org1MSP nar.blockr Orderer1MSP 0
+distribute_conf vm2 Org2MSP car.blockr Orderer2MSP 1
 
-rm -rf ./$CONFIG_DIR
+rm -rf $FABRIC_CFG_PATH
